20210419-Routine

+++ public/js/stores/items/show/categories/delete.js
@@ -1,20 +1,27 @@
-function categoryDeleteBtns() {
-    document.querySelectorAll('.categories').forEach(e => {
+function addCategoryDeleteBtn(category_id) {
+    let category_delete_btn = document.querySelector('#category_delete_btn');
+    if (category_delete_btn) {
+        category_delete_btn.innerHTML = '';
         get({
             table: 'item_category',
-            query: [`item_category_id=${e.dataset.id}`]
+            query: [`item_category_id=${category_id}`]
         })
-        .then(function([item_category, options]) {
-            e.appendChild(
+        .then(function ([category, options]) {
+            category_delete_btn.appendChild(
                 new Delete_Button({
+                    path: `/item_categories/${category.item_category_id}`,
                     descriptor: 'category',
-                    path:       `/item_categories/${item_category.item_category_id}`,
-                    small:      true,
-                    options:    {onComplete: getCategories}
+                    options: {
+                        onComplete: [
+                            getCategories,
+                            function () {$('#mdl_category_view').modal('hide')}
+                        ]
+                    }
                 }).e
             );
-            e.removeAttribute('data-id');
-            e.removeAttribute('class');
         });
-    });
+    };
 };
+window.addEventListener('load', function () {
+    $('#mdl_category_view').on('show.bs.modal', function (event) {addCategoryDeleteBtn(event.relatedTarget.dataset.id)})
+});
